+++ fwdpp/sugar/sampling/sampling_details.hpp
@@ -9,7 +9,7 @@ namespace KTfwd
   }
 
   template<typename vec_mutation_t>
-  void add_fixations( sample_t * sample,
+  void add_fixations( sample_t & sample,
 		      const vec_mutation_t & fixations,
 		      const unsigned nsam,
 		      const sugar::treat_neutral treat )
@@ -18,61 +18,70 @@ namespace KTfwd
       {
 	if( treat == sugar::treat_neutral::ALL )
 	  {
-	    sample->emplace_back( std::make_pair(f.pos,std::string(nsam,'1')) );
+	    sample.emplace_back( f.pos,std::string(nsam,'1') );
 	  }
 	else if (treat == sugar::treat_neutral::NEUTRAL && f.neutral ) //only add neutral mutations
 	  {
-	    sample->emplace_back( std::make_pair(f.pos,std::string(nsam,'1')) );
+	    sample.emplace_back( f.pos,std::string(nsam,'1') );
 	  }
 	else if (treat == sugar::treat_neutral::SELECTED && !f.neutral ) //only add selected mutations
 	  {
-	    sample->emplace_back( std::make_pair(f.pos,std::string(nsam,'1')) );
+	    sample.emplace_back( f.pos,std::string(nsam,'1') );
 	  }
       }
   }
 
+  template<typename vec_mutation_t>
+  void finish_sample( sample_t & sample, const vec_mutation_t & fixations, const unsigned nsam, const bool removeFixed, const sugar::treat_neutral treat )
+  {
+    if(! removeFixed )
+      {
+	add_fixations(sample,fixations,nsam,treat);
+      }
+    std::sort( sample.begin(),sample.end(),
+	       [](const sample_site_t & a,
+		  const sample_site_t & b) noexcept {
+		 return a.first<b.first;
+	       });
+  }
+  
   //Single-region, single-deme
   template<typename poptype>
-  sample_t sample_details( gsl_rng * r,
+  sample_t sample_details( const gsl_rng * r,
 			   const poptype & p,
 			   const unsigned nsam,
 			   const bool removeFixed,
 			   std::true_type)
   {
     sample_t rv =  ms_sample(r,p.mutations,p.gametes,p.diploids,nsam,removeFixed);
-    if(!removeFixed)
-      add_fixations(&rv,p.fixations,nsam,sugar::treat_neutral::ALL);
+    finish_sample(rv,p.fixations,nsam,removeFixed,sugar::treat_neutral::ALL);
     return rv;
   }
 
   //Multi-locus, single-deme
   template<typename poptype>
-  sample_t sample_details( gsl_rng * r,
-				   const poptype & p,
+  sample_t sample_details( const gsl_rng * r,
+			   const poptype & p,
 			   const unsigned nsam,
 			   const bool removeFixed,
 			   std::false_type)
   {
     sample_t rv = ms_sample(r,p.mutations,p.gametes,p.diploids,nsam,removeFixed);
-    if(!removeFixed)
-      add_fixations(&rv,p.fixations,nsam,sugar::treat_neutral::ALL);
+    finish_sample(rv,p.fixations,nsam,removeFixed,sugar::treat_neutral::ALL);
     return rv;
   }
 
   //Single-region, single-deme
   template<typename poptype>
-  sep_sample_t sample_sep_details( gsl_rng * r,
+  sep_sample_t sample_sep_details( const gsl_rng * r,
 				   const poptype & p,
 				   const unsigned nsam,
 				   const bool removeFixed,
 				   std::true_type)
   {
     sep_sample_t rv = ms_sample_separate(r,p.mutations,p.gametes,p.diploids,nsam,removeFixed);
-    if(! removeFixed)
-      {
-	add_fixations(&rv.first,p.fixations,nsam,sugar::treat_neutral::NEUTRAL);
-	add_fixations(&rv.second,p.fixations,nsam,sugar::treat_neutral::SELECTED);
-      }
+    finish_sample(rv.first,p.fixations,nsam,removeFixed,sugar::treat_neutral::NEUTRAL);
+    finish_sample(rv.second,p.fixations,nsam,removeFixed,sugar::treat_neutral::SELECTED);
     return rv;
   }
 
@@ -83,19 +92,9 @@ namespace KTfwd
 			   std::true_type)
   {
     sep_sample_t temp = fwdpp_internal::ms_sample_separate_single_deme(p.mutations,p.gametes,p.diploids,individuals,2*individuals.size(),removeFixed);
-    if(! removeFixed)
-      {
-	add_fixations(&temp.first,p.fixations,2*individuals.size(),sugar::treat_neutral::NEUTRAL);
-	add_fixations(&temp.second,p.fixations,2*individuals.size(),sugar::treat_neutral::SELECTED);
-      }
     auto rv = std::move(temp.first);
     std::move(temp.second.begin(),temp.second.end(),std::back_inserter(rv));
-    std::sort(rv.begin(),rv.end(),[](const sample_site_t & a,
-				     const sample_site_t & b){
-		return a.first<b.first;
-	      });
-    if(!removeFixed)
-      add_fixations(&rv,p.fixations,2*individuals.size(),sugar::treat_neutral::ALL);
+    finish_sample(rv,p.fixations,2*individuals.size(),removeFixed,sugar::treat_neutral::ALL);
     return rv;
   }
 
@@ -106,19 +105,9 @@ namespace KTfwd
 			   std::false_type)
   {
     sep_sample_t temp = fwdpp_internal::ms_sample_separate_single_deme(p.mutations,p.gamtes,p.diploids,individuals,2*individuals.size(),removeFixed);
-    if(! removeFixed)
-      {
-	add_fixations(&temp.first,p.fixations,2*individuals.size(),sugar::treat_neutral::NEUTRAL);
-	add_fixations(&temp.second,p.fixations,2*individuals.size(),sugar::treat_neutral::SELECTED);
-      }
     auto rv = std::move(temp.first);
     std::move(temp.second.begin(),temp.second.end(),std::back_inserter(rv));
-    std::sort(rv.begin(),rv.end(),[](const sample_site_t & a,
-				     const sample_site_t & b){
-		return a.first<b.first;
-	      });
-    if(!removeFixed)
-      add_fixations(&rv,p.fixations,2*individuals.size(),sugar::treat_neutral::ALL);
+    finish_sample(rv,p.fixations,2*individuals.size(),removeFixed,sugar::treat_neutral::ALL);
     return rv;
   }
   
@@ -129,28 +118,22 @@ namespace KTfwd
 				   std::true_type)
   {
     sep_sample_t rv = fwdpp_internal::ms_sample_separate_single_deme(p.mutations,p.gametes,p.diploids,individuals,2*individuals.size(),removeFixed);
-    if(! removeFixed)
-      {
-	add_fixations(&rv.first,p.fixations,2*individuals.size(),sugar::treat_neutral::NEUTRAL);
-	add_fixations(&rv.second,p.fixations,2*individuals.size(),sugar::treat_neutral::SELECTED);
-      }
+    finish_sample(rv.first,p.fixations,2*individuals.size(),removeFixed,sugar::treat_neutral::NEUTRAL);
+    finish_sample(rv.second,p.fixations,2*individuals.size(),removeFixed,sugar::treat_neutral::SELECTED);
     return rv;
   }
 
   //Multi-locus, single-deme
   template<typename poptype>
-  sep_sample_t sample_sep_details( gsl_rng * r,
+  sep_sample_t sample_sep_details( const gsl_rng * r,
 				   const poptype & p,
 				   const unsigned nsam,
 				   const bool removeFixed,
 				   std::false_type)
   {
     sep_sample_t rv =  ms_sample_separate(r,p.mutations,p.gametes,p.diploids,nsam,removeFixed);
-    if(! removeFixed)
-      {
-	add_fixations(&rv.first,p.fixations,nsam,sugar::treat_neutral::NEUTRAL);
-	add_fixations(&rv.second,p.fixations,nsam,sugar::treat_neutral::SELECTED);
-      }
+    finish_sample(rv.first,p.fixations,nsam,removeFixed,sugar::treat_neutral::NEUTRAL);
+    finish_sample(rv.second,p.fixations,nsam,removeFixed,sugar::treat_neutral::SELECTED);
     return rv;
   }
 }